Move FileSystemOperation, Copy and Link classes to file_system.py
