+++ tsrc/__init__.py
@@ -5,9 +5,10 @@ __version__ = "2.2.0"
 from .config import Config, parse_config  # noqa
 from .errors import Error, InvalidConfig  # noqa
 from .executor import Task, run_sequence, ExecutorFailed  # noqa
+from .file_system import Copy, Link, FileSystemOperation  # noqa
 from .groups import GroupList, Group  # noqa
 from .groups import GroupNotFound, UnknownElement as UnknownGroupElement  # noqa
-from .repo import Repo, Remote, Copy, Link, FileSystemOperation  # noqa
+from .repo import Repo, Remote  # noqa
 from .manifest import Manifest  # noqa
 from .manifest import load as load_manifest  # noqa
 from .workspace import Workspace  # noqa
